+++ src/lib/components/book-card/book-manager-header.svelte
@@ -485,7 +485,8 @@
                       mergeEntries.MANAGE,
                       mergeEntries.STATISTICS,
                       mergeEntries.NOTEBOOK,
-                      mergeEntries.SETTINGS
+                      mergeEntries.SETTINGS,
+                      mergeEntries.CHANGELOG
                     ]),
                 ...(isTauri() ? [mergeEntries.CHECK_UPDATE] : [])
               ]}